Fix test regressions on windows due to `Reset` never being called
I'm not sure this is great. Without calling `Reset`, the correct initial
time may not be set (ever). In practice this doesn't happen anywhere in
the gameplay flow, but something worth noting.

This change is required now that `Reset` isn't called in the
constructor. It couldn't be called in the constructor because it would
cause the audio track to reset its position too early.

What an ordeal.