Fix lead in tests not waiting for player to start running
The tests are only meant to ensure that gameplay eventually starts.

The case where failures can occur is where the master clock is behind
the player clock (due to being in lead-in time). Because the test is
running in real-time, it can take arbitrary amounts of time to catch up.

If it took too long, the test would fail.