Fix failing large offset test
If we are going to continue to let the underlying clock process frames,
there needs to be a bit of lenience to allow the backwards seek on
resume (to play back over the freq ramp period).

The test is meant to be ensuring we don't skip the full offset amount,
so div10 seems pretty safe.