Add osu! hit object dim
Stable would dim objects when they can't be hit (ie. the "miss" window
is not active yet). This was never implemented in lazer, and causes
quite large visual differences.

No one has mentioned this yet, but it will definitely be one of those
missing pieces which makes lazer feel different to stable.