Do a second pass over DrawableList.updateArrangement
This touches on general CodeQuality of commit 3e41e1c76befb3e8181aa4ca561a6be5ce62c983, and reworks a bit of stuff so it feels better.

This moves a part of DrawableList.updateArrangement into DrawableList.checkArrangement.
If no change needs to be made to the current layout or if the dragged element has been moved out of the list, we return a empty optional.
If the currentlyDraggedItem needs to be moved, a Value is returned.

This also removes the preprocessing statements that were added.