Applying mods did not update values in BeatmapAttributeText
Applying a mod with a difficulty change did previously not update the values displayed in BeatmapAttributeText.
This attempts to work around this, by taking heavy "inspiration" from `osu.Game.Screens.Select.Details.AdvancedStats` and `osu.Game.Screens.Select.BeatmapInfoWedge`.

Now the displayed mod statistics are the same in the Beatmap details section in the select screen, and in the values in BeatmapAttributeText.