Add `UserProfile` model to be used in user profile overlay
As `APIUser` implements `IEquatable`, attempting to replace an `APIUser`
with another `APIUser` with the same online ID has no effect on the user
profile overlay. This is a significant hurdle in implementing support
for viewing the profile for different rulesets, as in that case the
profile is basically reloaded for the same user, but slightly different
data.

To facilitate this, wrap `APIUser` in a new `UserProfile` class. This
will mean that the equality rules can be changed locally to the user
profile overlay without impacting other components that depend on the
`APIUser` equality rules. The ruleset that the user profile is being
displayed with will eventually be added to `UserProfile`, too.