Combine localisations for abstract skinnable components into a single file
Generally we don't want localisation files with only one to two
translations. It makes it harder for translators to handle in crowdin (a
lot of file changes for small results).

So for cases like this I believe we should be grouping translations
where it makes sense.

I've left individual components in their own files as I can see
potential for more settings to be added in the future. Plus it gives a
bit of extra context.