Fix settings jump buttons not always seeking to correctly location the first time
Having a tablet connected would cause the first jump to be too short if
the tablet settings have not come into view yet.

I don't know a better way to handle this. I'm also not sure how to best
guard against this happening in the future (without forcing `IsPresent`
for a frame or something, for all `SettingsSubsection`s) as it's a
pretty case-by-case implementation.

Closes https://github.com/ppy/osu/issues/23123.