Adjust `BlurredIcon` expansion ratio to fix glow getting cut off at lower resolutions
Closes https://github.com/ppy/osu/issues/23210.

Ballparked the fix to work down to the lowest resolution we support. The
previous number (`2.5f`) was also likely ballparked so the fix seems in
line with expectations.

I don't want to put too much thought into this because the design of
this screen is likely going to change in the mean time anyway.