Propagate samples to strong nested hits too
The rationale is the same as in
e478e0c029b94193ae00c7d48edbf3bd42ac3788. Due to the recursive nature of
`GameplaySampleTriggerSource.GetMostValidObject()`, in the case of
nested hits, drum rolls and drum roll ticks, the nested strong hits
would become the most valid object, and so without propagating
the samples down to that level too, nothing would play.