implement 3-button navigation for start of game
I mainly intended to use this myself to navigate osu with a taiko
controller, just starting out with the navigation screen here where it's
quite simple to design.

For taiko controller usage it's intended to have left/right rim for left
or right navigation and center drums as Select button. For a nicer
experience on arrow keys, up and down acts pretty much the same as
escape and enter respectively.

TODO: currently we reuse the hover effect for the focused main menu
buttons. The styling should be made a bit more visible for keyboard
navigation, since the hover isn't that visible afterwards if you didn't
see the animation.