Allow changing map conversion source ruleset
For custom rulesets that rather want to only base converted maps on
different rulesets, such as taiko, insteead of standard osu maps.

Previously when using a custom ruleset, all maps except osu! standard
maps would be hidden in the song list when auto-converting maps. Now
all maps except for the specified ruleset maps are hidden in the song
list when auto-converting maps.

This is the most simple thing I could think of to not hurt performance,
but ideally in the future we might want to change this API to support
multiple source rulesets, to show e.g. all osu! and taiko and mania maps
when auto-converting and not just a single source ruleset.