Adjust cyclic selection test to hitcircle lifetime adjustments
As it turns out, the tightening of hitcircle lifetimes in editor caused
the test to fail, since the objects were too far apart and at the
starting time of the test, the first object was fully faded out and as
such not alive, therefore leading cyclic selection to fail to select it.

To fix, bring all three objects closer together time-wise so that this
does not happen and the test can continue to exercise its original
behaviour.