AS mod: Use a power curve on the rate multiplier
The previous linear range would get clamped when the rate multiplier was around 0.90.

Using a power curve makes the setting's lower bound meaningful.

The default value should be 0.5085 to perfectly match previous replays on this branch, but 0.51 is imperceptibly close (rates differ by ~0.01), so we keep the bindable's precision as-is.

```
>>> min_allowable_rate_change / 0.95
0.9473684210526316
>>> pow(min_allowable_rate_change, 0.5085)
0.9478340714032744
>>> pow(min_allowable_rate_change, 0.51)
0.947684286809841

>>> max_allowable_rate_change * 0.95
1.0545
>>> pow(max_allowable_rate_change, 0.5085)
1.0545003657593834
>>> pow(max_allowable_rate_change, 0.51)
1.0546654501917019
```