Disable redundant verbatim string prefix inspections
For some reason this started flaring up recently all over for me and
showing inspections all over, which are _technically_ valid, but
interfere with our convention of using verbatim string prefixes to
denote non-localisable strings. This, as a result, led to circular
inspections (addressing the r# inspection results in getting the
osu-localisation-analyser one, addresssing that one results in
getting the r# inspection back, etc. ad nauseam).